Fix: corrigindo os tipos de atributos

+++ desafio-dio-banco/src/entities/ContaConrrente.java
@@ -4,7 +4,7 @@ public class ContaConrrente {
 
     private int agencia;
     private int contaCorrente;
-    private int saldo;
+    private double saldo;
 
     public void sacar(){
